refactor homepage and tab components
