+++ src/pages/Home.jsx
@@ -7,9 +7,9 @@ import { Context, populateHome } from "./../store";
 import { useParams } from "react-router-dom";
 /* widget/component imports */
 import DashboardContent from "../layouts/DashBoard";
-import TaskWithPartner from "../components/home/TaskWithPartner";
 import TaskWithNoPartner from "../components/home/TaskWithNoPartner";
-import TaskTest from "../components/home/TaskTest";
+import TaskWithPartner from "../components/home/TaskWithPartner";
+
 
 // Test date
 // import { parseJSON } from "date-fns"
@@ -48,8 +48,7 @@ function TabPanel({ children, value, index }) {
 
 export default function HomePage() {
   const { store, dispatch } = useContext(Context);
-  const { user, token, tasks } = store;
-  const [taskFiltering, setTaskFiltering] = useState([]);
+  const { user, token, tasks } = store;  
   const auth = { headers: { Authorization: `Bearer ${token}` } };
   const { userId } = useParams();
   let validId;
@@ -92,19 +91,10 @@ export default function HomePage() {
         )
       );
 
-      // filter and store different taskdata in home state for rendering
-      let taskStorage = [...tasks];
-      console.log("taskStorage", taskStorage);
-      let filteredTask = taskStorage.filter(
-        (task) => task.partner == null && task.completed === false
-      );
-      setTaskFiltering(filteredTask);
-
     })();
   }, []);
 
-  console.log("This is store in home.jsx", store);
-    console.log("This is taskFiltering", taskFiltering);
+    console.log("This is store in home.jsx", store);
 
   /**************
    * Function to change tabs
@@ -127,11 +117,14 @@ export default function HomePage() {
           </MyTabs>
         </Box>
         <TabPanel value={value} index={0}>
-          <TaskTest taskfiltering={taskFiltering} />
+          <TaskWithNoPartner />
         </TabPanel>
         <TabPanel value={value} index={1}>
-          {/* <TaskWithPartner /> */}
+          <TaskWithPartner />
         </TabPanel>
+        {/* <TabPanel value={value} index={2}>
+          <TaskTest />
+        </TabPanel> */}
       </Box>
     );
   }
@@ -142,7 +135,7 @@ export default function HomePage() {
 
   return (
     <DashboardContent>
-      {`This is token after signing in ${token}`}
+      {/* {`This is token after signing in ${token}`} */}
       <Grid item xs={12} sm={6} md={3}>
         <BasicTabs />
       </Grid>
